Export migrations to folder
