+++ app/Console/Kernel.php
@@ -31,6 +31,7 @@ class Kernel extends ConsoleKernel
         'App\Console\Commands\InitLookup',
         'App\Console\Commands\CalculatePayouts',
         'App\Console\Commands\UpdateKey',
+        'App\Console\Commands\ExportMigrations',
     ];
 
     /**
